+++ CHANGELOG.md
@@ -8,6 +8,7 @@
     * New: `jQuery.payment()` function to ease payment redirection.
     * Fix: Invalid formatting of shipping dropdown.
     * Fix: Displaying multiple select fields.
+    * Fix: Properly calculate tax for shipping.
 * 1.10.6 - 2014-07-30:
     * Fix: Security issue on comments feed.
     * Fix: Add obeying validate postcode setting in JavaScript validation.